Fix SelectField mess
